Cleaned up and integrated word list (Done, I think)
Removed code I used for testing and added a couple lines to turn the randomlist.txt into a Python list to work from. Putting it up here so I can get to it and test from my desktop if the randomlist works as well as my test list I was using.